move goto command to gm level 0
