+++ src/main/java/client/command/commands/gm0/GotoCommand.java
@@ -21,7 +21,7 @@
 /*
    @Author: Arthur L - Refactored command content into modules
 */
-package client.command.commands.gm1;
+package client.command.commands.gm0;
 
 import client.Character;
 import client.Client;
